cq_server/requirements.txt
setuptools
cffi
xxtea-py


pymysql
flask[async]
pyyaml
flask_cors
flask_socketio
sqlalchemy
flask_sqlalchemy
aiortc
aiohttp
eyed3
protobuf==3.19.0
requests
cryptography
lupa
python-telegram-bot
eventlet
rarfile
gevent
pyodbc
sqlacodegen
websocket-client
cryptography
resend
certbot
Pillow
qrcode
loguru
